Move integration tests to separate file
